New "small-scale" Tests
Add `small-scale/` subfolder in whitespace-bug with new tests to try and
isolate the WS bug by starting with a very small adventure and add new
constructs to it until the bug is replicated.

Also, the test adventure contain leading & trailing spaces, and the
test script creates a stripped copy and then compiles and tests both
versions with the same commands script to compare their output log for
white space differences.